third_party/WebKit/LayoutTests/fast/animation/request-animation-frame-expected.txt

Tests basic use of requestAnimationFrame

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS callbackInvoked is true
PASS successfullyParsed is true

TEST COMPLETE
